+++ app/javascript/dashboard/routes/dashboard/conversation/ContactPanel.vue
@@ -29,76 +29,13 @@
                 value => toggleSidebarUIState('is_conv_actions_open', value)
               "
             >
-              <div>
-                <div class="multiselect-wrap--small">
-                  <contact-details-item
-                    compact
-                    :title="$t('CONVERSATION_SIDEBAR.ASSIGNEE_LABEL')"
-                  >
-                    <template v-slot:button>
-                      <woot-button
-                        v-if="showSelfAssign"
-                        icon="ion-arrow-right-c"
-                        variant="link"
-                        size="small"
-                        @click="onSelfAssign"
-                      >
-                        {{ $t('CONVERSATION_SIDEBAR.SELF_ASSIGN') }}
-                      </woot-button>
-                    </template>
-                  </contact-details-item>
-                  <multiselect-dropdown
-                    :options="agentsList"
-                    :selected-item="assignedAgent"
-                    :multiselector-title="
-                      $t('AGENT_MGMT.MULTI_SELECTOR.TITLE.AGENT')
-                    "
-                    :multiselector-placeholder="
-                      $t('AGENT_MGMT.MULTI_SELECTOR.PLACEHOLDER')
-                    "
-                    :no-search-result="
-                      $t('AGENT_MGMT.MULTI_SELECTOR.SEARCH.NO_RESULTS.AGENT')
-                    "
-                    :input-placeholder="
-                      $t('AGENT_MGMT.MULTI_SELECTOR.SEARCH.PLACEHOLDER.AGENT')
-                    "
-                    @click="onClickAssignAgent"
-                  />
-                </div>
-                <div class="multiselect-wrap--small">
-                  <contact-details-item
-                    compact
-                    :title="$t('CONVERSATION_SIDEBAR.TEAM_LABEL')"
-                  />
-                  <multiselect-dropdown
-                    :options="teamsList"
-                    :selected-item="assignedTeam"
-                    :multiselector-title="
-                      $t('AGENT_MGMT.MULTI_SELECTOR.TITLE.TEAM')
-                    "
-                    :multiselector-placeholder="
-                      $t('AGENT_MGMT.MULTI_SELECTOR.PLACEHOLDER')
-                    "
-                    :no-search-result="
-                      $t('AGENT_MGMT.MULTI_SELECTOR.SEARCH.NO_RESULTS.TEAM')
-                    "
-                    :input-placeholder="
-                      $t('AGENT_MGMT.MULTI_SELECTOR.SEARCH.PLACEHOLDER.TEAM')
-                    "
-                    @click="onClickAssignTeam"
-                  />
-                </div>
-                <contact-details-item
-                  compact
-                  :title="
-                    $t('CONVERSATION_SIDEBAR.ACCORDION.CONVERSATION_LABELS')
-                  "
-                />
-                <conversation-labels :conversation-id="conversationId" />
-              </div>
+              <conversation-action
+                :conversation-id="conversationId"
+                :inbox-id="inboxId"
+              />
             </accordion-item>
           </div>
-          <div v-if="element.name === 'conversation_info'">
+          <div v-else-if="element.name === 'conversation_info'">
             <accordion-item
               :title="$t('CONVERSATION_SIDEBAR.ACCORDION.CONVERSATION_INFO')"
               :is-open="isContactSidebarItemOpen('is_conv_details_open')"
@@ -114,7 +51,7 @@
               </conversation-info>
             </accordion-item>
           </div>
-          <div v-if="element.name === 'contact_attributes'">
+          <div v-else-if="element.name === 'contact_attributes'">
             <accordion-item
               :title="$t('CONVERSATION_SIDEBAR.ACCORDION.CONTACT_ATTRIBUTES')"
               :is-open="isContactSidebarItemOpen('is_contact_attributes_open')"
@@ -136,7 +73,7 @@
               />
             </accordion-item>
           </div>
-          <div v-if="element.name === 'previous_conversation'">
+          <div v-else-if="element.name === 'previous_conversation'">
             <accordion-item
               v-if="contact.id"
               :title="
@@ -166,11 +103,10 @@ import agentMixin from '../../../mixins/agentMixin';
 
 import AccordionItem from 'dashboard/components/Accordion/AccordionItem';
 import ContactConversations from './ContactConversations.vue';
-import ContactDetailsItem from './ContactDetailsItem.vue';
+import ConversationAction from './ConversationAction.vue';
+
 import ContactInfo from './contact/ContactInfo';
 import ConversationInfo from './ConversationInfo';
-import ConversationLabels from './labels/LabelBox.vue';
-import MultiselectDropdown from 'shared/components/ui/MultiselectDropdown.vue';
 import CustomAttributes from './customAttributes/CustomAttributes.vue';
 import CustomAttributeSelector from './customAttributes/CustomAttributeSelector.vue';
 import draggable from 'vuedraggable';
@@ -180,13 +116,11 @@ export default {
   components: {
     AccordionItem,
     ContactConversations,
-    ContactDetailsItem,
     ContactInfo,
     ConversationInfo,
-    ConversationLabels,
-    MultiselectDropdown,
     CustomAttributes,
     CustomAttributeSelector,
+    ConversationAction,
     draggable,
   },
   mixins: [alertMixin, agentMixin, uiSettingsMixin],
@@ -242,61 +176,6 @@ export default {
       const { custom_attributes: customAttributes } = this.contact;
       return customAttributes && Object.keys(customAttributes).length;
     },
-    teamsList() {
-      if (this.assignedTeam) {
-        return [
-          {
-            id: 0,
-            name: 'None',
-          },
-          ...this.teams,
-        ];
-      }
-      return this.teams;
-    },
-    assignedAgent: {
-      get() {
-        return this.currentChat.meta.assignee;
-      },
-      set(agent) {
-        const agentId = agent ? agent.id : 0;
-        this.$store.dispatch('setCurrentChatAssignee', agent);
-        this.$store
-          .dispatch('assignAgent', {
-            conversationId: this.currentChat.id,
-            agentId,
-          })
-          .then(() => {
-            this.showAlert(this.$t('CONVERSATION.CHANGE_AGENT'));
-          });
-      },
-    },
-    assignedTeam: {
-      get() {
-        return this.currentChat.meta.team;
-      },
-      set(team) {
-        const teamId = team ? team.id : 0;
-        this.$store.dispatch('setCurrentChatTeam', team);
-        this.$store
-          .dispatch('assignTeam', {
-            conversationId: this.currentChat.id,
-            teamId,
-          })
-          .then(() => {
-            this.showAlert(this.$t('CONVERSATION.CHANGE_TEAM'));
-          });
-      },
-    },
-    showSelfAssign() {
-      if (!this.assignedAgent) {
-        return true;
-      }
-      if (this.assignedAgent.id !== this.currentUser.id) {
-        return true;
-      }
-      return false;
-    },
   },
   watch: {
     conversationId(newConversationId, prevConversationId) {
@@ -330,44 +209,7 @@ export default {
     openTranscriptModal() {
       this.showTranscriptModal = true;
     },
-    onSelfAssign() {
-      const {
-        account_id,
-        availability_status,
-        available_name,
-        email,
-        id,
-        name,
-        role,
-        avatar_url,
-      } = this.currentUser;
-      const selfAssign = {
-        account_id,
-        availability_status,
-        available_name,
-        email,
-        id,
-        name,
-        role,
-        thumbnail: avatar_url,
-      };
-      this.assignedAgent = selfAssign;
-    },
-    onClickAssignAgent(selectedItem) {
-      if (this.assignedAgent && this.assignedAgent.id === selectedItem.id) {
-        this.assignedAgent = null;
-      } else {
-        this.assignedAgent = selectedItem;
-      }
-    },
 
-    onClickAssignTeam(selectedItemTeam) {
-      if (this.assignedTeam && this.assignedTeam.id === selectedItemTeam.id) {
-        this.assignedTeam = null;
-      } else {
-        this.assignedTeam = selectedItemTeam;
-      }
-    },
     onDragEnd() {
       this.dragging = false;
       this.updateUISettings({
